refactor: extract raw_output envelope encoding into pg_raw_output module
