+++ backend/windmill-worker/src/lib.rs
@@ -59,6 +59,7 @@ mod otel_oss;
 mod otel_tracing_proxy_ee;
 mod otel_tracing_proxy_oss;
 pub mod pg_executor;
+mod pg_raw_output;
 #[cfg(feature = "php")]
 mod php_executor;
 mod prepare_deps;
